ai-service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
google-generativeai==0.3.2
groq==0.4.2
python-dotenv==1.0.0
httpx==0.26.0
